+++ server/server-agent.ts
@@ -134,6 +134,7 @@ export namespace ServerAgent {
       // even not on verbose mode.
       if (this.logLevel !== 0) {
         console.log(`pageUrl: [${pageUrl}]`);
+        console.log(`serverless pageUrl: [${this.url}]`);
       }
       return pageUrl;
     }