+++ shared/dev-builders/build-docs.js
@@ -41,6 +41,7 @@ var BuildDocs;
     var MARKDOWN_FOLDER = 'docs';
     BuildDocs.API_FOLDER = 'api';
     BuildDocs.EN_LAST_VERSION_PATH = 'versions/latest/en';
+    var badgeLine = '';
     BuildDocs.targets = [
         {
             id: 'end-user',
@@ -51,6 +52,12 @@ var BuildDocs;
             indexFile: './README.md',
             isEndUser: true,
             logFile: './build-docs-end-user.log',
+            processIndexPage: function (data) {
+                return data.replace(/^(.*)developer-badge\.gif(.*)$/m, function (all, p1, p2) {
+                    badgeLine = all;
+                    return p1 + 'end-user-badge.gif' + p2;
+                });
+            },
         },
         {
             id: 'dev',
@@ -61,6 +68,15 @@ var BuildDocs;
             indexFile: dev_paths_js_1.DevPaths.SOURCE_DOCS_PATH + "-dev/README.md",
             isEndUser: false,
             logFile: './build-docs-dev.log',
+            processIndexPage: function (data) {
+                return data.replace(/^(# Description.*)$/m, function (all, p1, p2) {
+                    if (!badgeLine) {
+                        throw "end-user should had been processed already.";
+                    }
+                    console.warn('---WARN: Make sure the README links match mkdocs-local---');
+                    return all + '\n' + badgeLine + '  \n';
+                });
+            },
         },
     ];
     // ------------------------------------------------------------------------
@@ -535,9 +551,12 @@ var BuildDocs;
      * Copies a markdown file from to the destination
      * and adds the file to mkdocs.
      */
-    function copyMarkdownFile(srcFileName, dstFileName, mkDocs, mkDocsOpts) {
+    function copyMarkdownFile(srcFileName, dstFileName, mkDocs, mkDocsOpts, processPage) {
         var data = fsix_js_1.fsix.readUtf8Sync(srcFileName);
         data = mkDocs.addSourceFile(mkDocsOpts, dstFileName, data);
+        if (processPage) {
+            data = processPage(data);
+        }
         sysFs.writeFileSync(dstFileName, data);
         return dstFileName;
     }
@@ -561,7 +580,7 @@ var BuildDocs;
                 refs: {},
             };
             // index.html
-            copyMarkdownFile(target.indexFile, markdownDstPath + "/index.md", mkDocsYml, {});
+            copyMarkdownFile(target.indexFile, markdownDstPath + "/index.md", mkDocsYml, {}, target.processIndexPage);
             // copy sources
             target.sourcePaths.forEach(function (sourcesPathName) {
                 sysFs.readdirSync(sourcesPathName).forEach(function (file) {