+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.8] -
+### Fixed
+- `_initBrowser();` is only executed during `story.constructor` in order to work with JSFiddle
+
 ## [0.2.7] -
 ### Changed
 - Relaxes the `abeamer create` to support a project names with spaces and non-latin characters.